This is a Jupyter notebook of the final project.
Skills Network Labs are undergoing maintenance to improve the security of your exercises. This is not directly related to CV Studio, but unfortunately affects it.

We are actively working on restoring the labs but it may take a few days for the lab systems to be fully restored. 

Once it will get fixed we'll let you know.